small but big update git ignore
I had to delete master branche bec if i compile the bot with the token:
1) token gets diactivated bec discord searches for published tokens
2) The token can be read in the .jar file
I fixed the problem by adding .jar and .txt in a git ignore file.
So i can have the current token in a txt file to work with AND i dont have to think about deleting compiled java code and breaking everything.